Created LabelMenu component and moved all necesarry code from cardcomponent. Passing task's label states through props to labelmenu
